backend/requirements.txt
fastapi
uvicorn 
python-dotenv 
requests
ollama
groq
python-jose[cryptography] 
bcrypt
python-multipart
psycopg2-binary 
sqlalchemy
alembic
pydantic[email]